+++ packages/gatsby-source-store/src/index.ts
@@ -1,1 +1,11 @@
-export * as GatsbyNode from './gatsby-node'
+import type { GraphQLSchema } from 'graphql'
+
+export { createSchemaCustomization } from './customizeSchema'
+export { sourceProducts } from './sourceProducts'
+export { sourceCollections } from './sourceCollections'
+
+export interface Options {
+  maxNumItems?: number
+  getSchema: () => Promise<GraphQLSchema>
+  getContextFactory: () => Promise<any>
+}
